move generator.py to tests
fix on 18.04:
Traceback (most recent call last):
  File ".../tests/python/test_Tasks.py", line 268, in <module>
    from models.arm.generator import create_7dof_arm
ImportError: No module named models.arm.generator
